fix: close all 17 SonarCloud findings and clarify coverage gating
All 17 issues flagged on PR #259 by SonarCloud were real (no false
positives): java:S1192 duplicated string literals in the snapshot and
fingerprint codecs, one unnecessary long cast, and 10 java:S5778
assertThrows lambdas performing more than one invocation that could
throw. Extract the duplicated literals into named constants and hoist
the extra invocations (values()[0], repeat(), chained getters) out of
the asserted lambdas so each assertThrows exercises exactly one call.

Also document why SonarCloud reports 0.0% coverage on new code: the
project runs in Automatic Analysis (GitHub App) mode, which cannot
import JaCoCo XML by design, regardless of build-side JaCoCo config.
Real coverage enforcement for new code is, and remains, the targeted
JaCoCo gate in PR Validation. Add an opt-in, off-by-default CI-based
Sonar analysis step (gated on a repository variable) so switching to
CI-based analysis - once SONAR_TOKEN is provisioned and Automatic
Analysis is disabled in SonarCloud project settings, both external
actions - is a one-line flip instead of a workflow rewrite.

Tests: full reactor test suite for minos-domain, minos-nexus,
minos-storage-local and minos-application passes unchanged (300 + 10
tests). No production behavior changes outside the two fixed
call-site constants.

+++ minos-application/src/main/java/com/minos/incremental/ProjectFingerprint.java
@@ -20,7 +20,7 @@ public record ProjectFingerprint(
         projectSha256 = FileFingerprint.requireSha256(projectSha256);
         buildSha256 = FileFingerprint.requireSha256(buildSha256);
         Objects.requireNonNull(files, "files");
-        if ((long) files.size() > SourceBudgetPolicy.DEFAULT_MAX_FILES) {
+        if (files.size() > SourceBudgetPolicy.DEFAULT_MAX_FILES) {
             throw new IllegalArgumentException(
                     "files exceeds source budget: " + files.size() + "/" + SourceBudgetPolicy.DEFAULT_MAX_FILES);
         }
